Modify parsing of optional fields
Logic for the resetting of optional fields such as "link/" and "sal/" in Person has been shifted from ParserUtil.java to EditPersonCommand.java. This is so that other commands (apart from "edit can") will not be affected by it as this might cause bugs. These optional fields can only be reset in "edit can" commands so the logic should be localised in the EditPersonCommand class, not in the ParserUtil class used by multiple commands.